refactor: merge options into respective menus
